refactor: split contracts and wallets logic in blocks and logs
